+++ src/astronomo/widgets/certificates/__init__.py
@@ -3,6 +3,8 @@
 from astronomo.widgets.certificates.confirm_delete_modal import ConfirmDeleteModal
 from astronomo.widgets.certificates.create_modal import CreateIdentityModal
 from astronomo.widgets.certificates.edit_modal import EditIdentityModal
+from astronomo.widgets.certificates.file_picker_modal import FilePickerModal
+from astronomo.widgets.certificates.import_custom_modal import ImportCustomModal
 from astronomo.widgets.certificates.import_lagrange_modal import ImportLagrangeModal
 from astronomo.widgets.certificates.manage_urls_modal import ManageUrlsModal
 
@@ -10,6 +12,8 @@ __all__ = [
     "ConfirmDeleteModal",
     "CreateIdentityModal",
     "EditIdentityModal",
+    "FilePickerModal",
+    "ImportCustomModal",
     "ImportLagrangeModal",
     "ManageUrlsModal",
 ]